add a cast between size_t and unsigned int
git-svn-id: svn://scm.gforge.inria.fr/svnroot/paradiseo@1222 331e1502-861f-0410-8da2-ba01fb791d7f

+++ trunk/paradiseo-moeo/src/diversity/moeoNearestNeighborDiversityAssignment.h
@@ -163,7 +163,7 @@ private:
     double getElement(std::list<double> _myList)
     {
         std::list<double>::iterator it= _myList.begin();
-        for (unsigned int i=1; i< std::min(_myList.size(),index); i++)
+        for (unsigned int i=1; i< std::min((unsigned int)_myList.size(),index); i++)
             it++;
         return *it;
     }
